+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "de.railslove.wfdcheckin",
+    package = "de.railslove.meinecheckins",
 )
 
 android_resource(
     name = "res",
-    package = "de.railslove.wfdcheckin",
+    package = "de.railslove.meinecheckins",
     res = "src/main/res",
 )
