+++ android/app/build.gradle
@@ -131,7 +131,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "de.railslove.wfdcheckin"
+        applicationId "de.railslove.meinecheckins"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 38
