+++ android/app/src/main/java/de/railslove/meinecheckins/MainApplication.java
@@ -1,4 +1,4 @@
-package de.railslove.wfdcheckin;
+package de.railslove.meinecheckins;
 
 import android.app.Application;
 import android.content.Context;
@@ -62,7 +62,7 @@ public class MainApplication extends Application implements ReactApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("de.railslove.wfdcheckin.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("de.railslove.meinecheckins.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
